+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+* 2016-01-03 Added support for inheritance and composition.
 * 2015-12-21 Added support for nested templates (layouts/masters).
 * 2015-12-21 Fixed an issue with templates being duplicated on synchronization.
 * 2015-12-09 Compiled using Umbraco 7.3.4 binaries.